backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.3
python-multipart==0.0.20
bcrypt==4.0.1
python-dotenv==1.0.1
email-validator==2.2.0
